requirements.txt
certifi==2017.7.27.1
chardet==3.0.4
future==0.16.0
idna==2.6
python-telegram-bot==7.0.1
requests==2.18.4
urllib3==1.22
